Combine sahred tag table methods into common trait
